fixed format once again
git-svn-id: https://bugengine.googlecode.com/svn/trunk@349 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/plugin/win32/src/renderer.cc
@@ -133,7 +133,7 @@ UINT Renderer::messageCount() const
 
 void Renderer::postMessage(UINT msg, WPARAM wParam, LPARAM lParam) const
 {
-    PostThreadMessageA(m_windowManagementThread.id(), msg, wParam, lParam);
+    PostThreadMessageA((DWORD)m_windowManagementThread.id(), msg, wParam, lParam);
 }
 
 void Renderer::handleMessage(UINT msg, WPARAM wParam, LPARAM lParam)
